InfoBubble/Description: move panel to its own view component
